Pre-compile asset files
Even after drastically improving the performance of the Sunder compiler,
compiling and linking asset file data created from `embed` expressions
was still dominating the compilation time ofr the project. Prior to this
commit, a release build of Natac took ~6 seconds on my M2 Mac, and ~5 of
those seconds were spent in the C compiler.

This changeset introduces an asset bundler, `misc/bundle-assets.sunder`,
which generates a C translation unit and Sunder module that can be used
to pre-compile asset files into a .o file, the contents of which can be
externally referenced from Sunder sources without needing to `embed` and
re-compile the multi-MB asset files every time the application is built.
Project compilation times after the initial build of the asset bundle
are now ~1.8 seconds for a debug build and ~3.2 seconds for a release
build on my M2 Mac.