Revert "Pre-compile asset files"
This reverts commit 6446908599f0267ba98df7fb40bf1cf67d4727f3.

For some reason, the offending commit caused raylib to segfault when
loading assets from memory, even though the pre-compiled asset data was
byte-for-byte identical to the assets on disk (verified by reading the
entire asset file off disk and comparing it byte-for-byte to the
pre-compiled data). I have absolutely no idea what is causing the bad
memory access, and I do not have the spoons to debug it properly right
now, so I am reverting the offending commit for the time being.